Crate shared_display_core

shared-display-core contains the core components of [shared-display] that are required to add screen-sharing support for a display driver.

This crate should only be used by drivers that extend shared-display. Applications should instead depend on [shared-display] itself.

This crate heavily relies on and builds on top of embedded-graphics and various crates of the embassy project.

Structs

CompressedBuffer

An RLE-encoded framebuffer.

CompressedDisplayPartition

A partition of a CompressableDisplay.

DecompressingIter

A decompressing Iterator for an RLE-encoded CompressedBuffer.

DisplayPartition

A partition of a SharableBufferedDisplay.

DrawTracker

An object to track the AreaToFlush in a concurrent context. Provides safe methods to read and write concurrently.

FlushLock

A lock to avoid writes to the buffer during decompression for flushing, but allow multiple writes at the same time.

Enums

AppEvent

Events from other apps that allow to alter a partition.

AreaToFlush

Area of the screen that has been drawn to since the last flush.

EnvelopeError

Things that might go wrong trying to envelope the area of an app that closed.

NewPartitionError

Error Type for creating new screen partitions.

Constants

MAX_APPS_PER_SCREEN

Maximum number of apps allowed on the screen concurrently.

Traits

CompressableDisplay

A SharableBufferedDisplay that can compressed.

SharableBufferedDisplay

A buffered DrawTarget that can be shared among multiple apps.